Use Arrow type system directly in Table, reduced-memory json output (#540)
* switch to arrow type system, some failing tests

* tests pass

* cleanup

* integrated comments, squashed some switched, clarified the breaking changes to GroupByOp

* format

* cleanup

* json output is now row-wise like neo4j, version bump on inputs, addressed nits

* json verification update for new path structure

* updated sum for new type system, added json parameter for graph-query, disabled failing distributed sum() test

* format

* g++ is picky, is this better?

* format

* g++ doesn't like getElement, it's private now

[REWRITTEN COMMIT: incomplete subtree]

+++ scripts/json_checker.py
@@ -9,25 +9,53 @@ import argparse
 import os
 import json
 import sys
+import logging
 
-def main(resultjson, truthdir):
-  failures = 0
+def compare(resultfile, truthfile):
+  try:
+    with open(resultfile) as result:
+      result_lines = [ line.strip() for line in result ]
+    with open(truthfile) as truth:
+      truth_lines = [ line.strip() for line in truth ]
+    if len(result_lines) != len(truth_lines):
+      logging.error("length mismatch: result=%s truth=%s", len(result_lines), len(truth_lines))
+      return False
+    for r,t in zip(result_lines, truth_lines):
+      jr = json.loads(r)
+      jt = json.loads(t)
+      if jr != jt:
+        logging.error("mismatch: %s VS %s", jr, jt)
+        return False
+    return True
+  except FileNotFoundError:
+    logging.error("a file is missing!")
+    return False
+  except Exception as e:
+    logging.error("exception: %s", str(e))
+    return False
+
+def main(querylist, resultprefix, truthdir, delete):
   successes = 0
-  with open(resultjson) as results_file:
-    results = json.load(results_file)
-    for query in results:
-      with open(truthdir + "/" + query + ".truth") as truth_file:
-        truth = json.load(truth_file)
-        if results[query] != truth:
-          print('Error at', query)
-          failures += 1
-        else:
-          successes += 1
+  failures = 0
+  with open(querylist) as queries_file:
+    for query_name in queries_file:
+      query_name = query_name.strip()
+      query_filename = resultprefix + "_" + query_name + ".json"
+      truth_filename = truthdir + "/" + query_name + ".truth"
+      ret = compare(query_filename, truth_filename)
+      if ret:
+        successes += 1
+        if delete:
+          os.remove(query_filename)
+      else:
+        failures += 1
+        logging.error('mismatch at query %s', query_name)
   if failures > 0:
     print("\nFAILED (", failures, "/", (successes+failures), ")\n")
+    return 1
   else:
     print("\nSUCCESS (", successes, ")\n")
-  return failures
+    return 0
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Check graph output results as json")
@@ -41,8 +69,12 @@ if __name__ == "__main__":
         help="delete the generated output files",
     )
     parser.add_argument(
-        "-resultjson",
-        help="file containing json of results"
+        "-querylist",
+        help="list of queries being tested"
+    )
+    parser.add_argument(
+        "-resultprefix",
+        help="prefix for files containing json of results"
     )
     parser.add_argument(
         "-truthdir",
@@ -52,11 +84,12 @@ if __name__ == "__main__":
     parsed_arguments = parser.parse_args()
 
     print("Starting comparison...")
-    ret = main(parsed_arguments.resultjson, parsed_arguments.truthdir)
+    ret = main(parsed_arguments.querylist, parsed_arguments.resultprefix,
+               parsed_arguments.truthdir, parsed_arguments.delete)
     if parsed_arguments.delete:
-        os.remove(parsed_arguments.resultjson)
-        countfile = parsed_arguments.resultjson.replace('.json', '.count')
-        os.remove(countfile)
+        countfile = parsed_arguments.resultprefix + ".count"
+        if os.path.exists(countfile):
+          os.remove(countfile)
 
     if ret:
         sys.exit(1)
